c301: extract StabilityTab cluster (KSensitivityPanel + LADDER_METHODS + ladderColor + StabilityLadderPanel + StabilityMatrixView + StabilityTab + StabilityTabBody + SceneStabilityStat) (#441 2.1)
Workspace.tsx 7362 LOC -> 6656 LOC (-706 LOC). Entry chunk 186.56 KB -> 171.96 KB.
20 tabs extracted + lazy. Cumulative since c262: -45%.
Removed unused useQueries + StabilityHeatmap imports from Workspace.tsx.
typecheck clean, 44/44 tests pass.
